+++ crates/arf-console/src/config/mod.rs
@@ -12,8 +12,7 @@ mod startup;
 pub use colors::{ColorsConfig, MetaColorConfig, RColorConfig, StatusColorConfig, ViColorConfig};
 pub use completion::CompletionConfig;
 pub use editor::{AutoSuggestions, EditorConfig, EditorMode};
-pub use experimental::SpinnerConfig;
-pub use experimental::{ExperimentalConfig, HistoryForgetConfig};
+pub use experimental::{ElapsedConfig, ExperimentalConfig, HistoryForgetConfig, SpinnerConfig};
 pub use history::HistoryConfig;
 pub use mode::ModeConfig;
 #[allow(unused_imports)]